Extract perl-dap debugger validation microcrate
