+++ crates/perl-dap-debugger/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "perl-dap-debugger"
+version = "0.11.0"
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+description = "Shared debugger parsing and validation utilities for perl-dap"
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+documentation = "https://docs.rs/perl-dap-debugger"
+readme = "README.md"
+keywords = ["perl", "dap", "debugger"]
+categories = ["development-tools::debugging"]
+include = ["src/**", "Cargo.toml", "README.md", "LICENSE*"]
+
+[lib]
+name = "perl_dap_debugger"
+path = "src/lib.rs"
+
+[dependencies]
+regex = "1.12.2"
+
+[lints]
+workspace = true
