+++ crates/perl-dap-debugger/README.md
@@ -1,0 +1,3 @@
+# perl-dap-debugger
+
+Shared debugger parsing and validation utilities extracted from `perl-dap` to keep the DAP adapter focused on transport and session orchestration.
