+++ crates/perl-dap/Cargo.toml
@@ -57,6 +57,7 @@ perl-dap-platform = { workspace = true }
 perl-dap-command-args = { workspace = true }
 perl-dap-variables = { workspace = true }
 perl-dap-stack = { workspace = true }
+perl-dap-debugger = { workspace = true }
 perl-module-path = { workspace = true }
 perl-dap-security = { workspace = true }
 perl-content-length-framing = { workspace = true }
