+++ crates/perl-dap/src/debug_adapter/execution.rs
@@ -1,6 +1,7 @@
 //! Execution control: continue, next, step in, step out, pause, goto, cancel.
 
 use super::*;
+use regex::Regex;
 
 impl DebugAdapter {
     /// Handle continue request
